requirements.txt
virustotal-api
qrcode
PySide2
Image